Change OnTeamsMessagingExtensionFetchTaskAsync to take MessagingExtensionAction and not MessagingExtensionQuery.
Change bad preview edit/send and bad file consent to return BadRequest.
